Combined code for signin and authentication, course selection and UI.
The commit combines all the features that each member was working on. This includes, the login/signin page and authentication using Firebase, course selection by retriving data from Firestore, and frontend UI for course selection page and dashboard. At this commit, developer could use a demo user to load the courses from Firestore, and reflect that data on to our webpage. Thecourses shown in the dashboard page is stil hardcoded since we don't have the algorithm that will allow us to make suggestion s for future courses. User is able to navigate through different pages, using 'react-router-dom's useNavigate. This version has not merged the additional email verification that Phi implemented as well as functions that alex added.
