--- react_dars_website/src/App.css
@@ -1,11 +1,0 @@
-h1 {
-  /* font-size: 4rem;
-  font-family: 'Abril Fatface';
-  text-transform: uppercase;
-  text-decoration: underline;
-  text-decoration-color: #7223d8;
-  text-decoration-thickness: 0.4rem;
-  text-underline-offset: 0.4rem;
-  color: #f9f9f9;
-  margin-top: 80px; */
-}
