+++ react_dars_website/src/Components/Pages/dashboard/DashBoard.js
@@ -12,7 +12,6 @@ function DashBoard(){
     return (
         <div className="">        
             <Navigationbar />
-            {/* <Example /> */}
             <Classfield />
 
         </div>
