+++ react_dars_website/src/Components/Pages/loginPage/ClassData.js
@@ -1,0 +1,335 @@
+export const ClassData = [
+    {
+        className: 'Intro to Comp Sci I',
+        classCode: 'CS 31',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: 'Intro to Comp Sci II',
+        classCode: 'CS 32',
+        availability: true,
+        prereq: ['CS 31'],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: 'Intro to Comp Organization',
+        classCode: 'CS 33',
+        availability: true,
+        prereq: ['CS 32'],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: 'Software Construction',
+        classCode: 'CS 35L',
+        availability: true,
+        prereq: ['CS 31'],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: 'Logic Design of Digital Systems',
+        classCode: 'CS M51A or ECE M16',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: 'Differential and Inegral Calculus',
+        classCode: 'Math 31A',
+        availability: true,
+        prereq: ['CS 31'],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: 'Integration and Infinite Series',
+        classCode: 'Math 31B',
+        availability: true,
+        prereq: ['Math 31A'],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: 'Integration and Infinite Series',
+        classCode: 'Math 31B',
+        availability: true,
+        prereq: ['Math 31A'],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: 'Calculus of Several Variables',
+        classCode: 'Math 32A',
+        availability: true,
+        prereq: ['Math 31A'],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: 'Calculus of Several Variables',
+        classCode: 'Math 32B',
+        availability: true,
+        prereq: ['Math 31B', 'Math 32A'],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: 'Integration and Infinite Series',
+        classCode: 'Math 33A',
+        availability: true,
+        prereq: ['Math 31B'],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'Math 33B',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'Math 61',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'Physics 1A',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'Physics 1B',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'Physics 1C',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'Physics 4AL or 4BL',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS 111',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS 118',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS 131',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS M151B or ECE M116C',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS M152A or ECE M116L',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS 180',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS 181',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS 130 or 152B',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'Statisitics (Stats 100A, Math 170A, Math 170E ECE 131A, CEE 110)',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS Upper Div Elective #1 (CS 100-188)',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS Upper Div Elective #2 (CS 100-188)',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS Upper Div Elective #3 (CS 100-187)',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS Upper Div Elective #4 (CS 100-187)',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'CS Upper Div Elective #5 (CS 100-187)',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'TBR #1',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'TBR #2',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'TBR #3',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'Sci-Tech #1',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'Sci-Tech #2',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    {
+        className: '',
+        classCode: 'Sci-Tech #3',
+        availability: true,
+        prereq: [],
+        completed: false,
+        grade: 0,
+        units: 3,
+    },
+    ];
