+++ react_dars_website/src/Components/Pages/selectionPage/SelectionCard.css
@@ -1,46 +1,29 @@
-@import url("https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700;800;900&display=swap");
-
-*,
-*:after,
-*:before {
-	box-sizing: border-box;
-}
-
-/* body {
-	font-family: "Inter", sans-serif;
-	line-height: 1.5;
-	min-height: 100vh;
-	display: flex;
-    margin-right: 5rem;
-	align-items: auto;
-	justify-content: flex-start;
-	background-color: #ffffff;
-} */
-
 .checkbox-group {
 	display: flex;
 	flex-wrap: wrap;
-	justify-content: flex-start;
+	justify-content: center;
 	width: 90%;
-	max-width: 90%;
+	margin-left: auto;
+	margin-right: auto;
+	max-width: 600px;
 	user-select: none;
 	& > * {
-		margin: .5rem 0.5rem;
+	  margin: .5rem 0.5rem;
 	}
-}
-
-
-
-.checkbox-group-legend {
+  }
+  
+  
+  
+  .checkbox-group-legend {
 	font-size: 1.5rem;
 	font-weight: 700;
 	color: #9c9c9c;
-	text-align: start;
+	text-align: center;
 	line-height: 1.125;
 	margin-bottom: 1.25rem;
-}
-
-.checkbox-input {
+  }
+  
+  .checkbox-input {
 	clip: rect(0 0 0 0);
 	clip-path: inset(100%);
 	height: 1px;
@@ -48,89 +31,90 @@
 	position: absolute;
 	white-space: nowrap;
 	width: 1px;
-
+  
 	&:checked + .checkbox-tile {
+	  border-color: #0A93DC;
+	  box-shadow: 0 5px 10px rgba(#000, 0.1);
+	  color: #0A93DC;
+	  &:before {
+		transform: scale(1);
+		opacity: 1;
+		background-color: #0A93DC;
 		border-color: #0A93DC;
-		box-shadow: 0 5px 10px rgba(#000, 0.1);
+	  }
+	  
+	  .checkbox-icon, .checkbox-label {
 		color: #0A93DC;
-		&:before {
-			transform: scale(1);
-			opacity: 1;
-			background-color: #0A93DC;
-			border-color: #0A93DC;
-		}
-		
-		.checkbox-icon, .checkbox-label {
-			color: #0A93DC;
-		}
+	  }
 	}
 	
 	&:focus + .checkbox-tile {
-		border-color: #0A93DC;
-		box-shadow: 0 5px 10px rgba(#000, 0.1), 0 0 0 4px #b5c9fc;
-		&:before {
-			transform: scale(1);
-			opacity: 1;
-		}
+	  border-color: #0A93DC;
+	  box-shadow: 0 5px 10px rgba(#000, 0.1), 0 0 0 4px #b5c9fc;
+	  &:before {
+		transform: scale(1);
+		opacity: 1;
+	  }
 	}
-}
-
-.checkbox-tile {
+  }
+  
+  .checkbox-tile {
 	display: flex;
 	flex-direction: column;
 	align-items: center;
 	justify-content: center;
-	width: 8rem;
-	min-height: 4rem;
-	border-radius: 0.375rem;
-	border: 1px solid #a1a1aa;
+	width: 16rem;
+	min-height: 8rem;
+	border-radius: 0.5rem;
+	border: 2px solid #b5bfd9;
 	background-color: #fff;
 	box-shadow: 0 5px 10px rgba(#000, 0.1);
 	transition: 0.15s ease;
 	cursor: pointer;
 	position: relative;
-
+  
 	&:before {
-		content: "";
-		position: absolute;
-		display: block;
-		width: 1.25rem;
-		height: 1.25rem;
-		border: 2px solid #b5bfd9;
-		background-color: #fff;
-		border-radius: 50%;
-		top: 0.25rem;
-		left: 0.25rem;
-		opacity: 0;
-		transform: scale(0);
-		transition: 0.25s ease;
-		background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='192' height='192' fill='%23FFFFFF' viewBox='0 0 256 256'%3E%3Crect width='256' height='256' fill='none'%3E%3C/rect%3E%3Cpolyline points='216 72.005 104 184 48 128.005' fill='none' stroke='%23FFFFFF' stroke-linecap='round' stroke-linejoin='round' stroke-width='32'%3E%3C/polyline%3E%3C/svg%3E");
-		background-size: 12px;
-		background-repeat: no-repeat;
-		background-position: 50% 50%;
+	  content: "";
+	  position: absolute;
+	  display: block;
+	  width: 1.25rem;
+	  height: 1.25rem;
+	  border: 2px solid #b5bfd9;
+	  background-color: #fff;
+	  border-radius: 50%;
+	  top: 0.25rem;
+	  left: 0.25rem;
+	  opacity: 0;
+	  transform: scale(0);
+	  transition: 0.25s ease;
+	  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='192' height='192' fill='%23FFFFFF' viewBox='0 0 256 256'%3E%3Crect width='256' height='256' fill='none'%3E%3C/rect%3E%3Cpolyline points='216 72.005 104 184 48 128.005' fill='none' stroke='%23FFFFFF' stroke-linecap='round' stroke-linejoin='round' stroke-width='32'%3E%3C/polyline%3E%3C/svg%3E");
+	  background-size: 12px;
+	  background-repeat: no-repeat;
+	  background-position: 50% 50%;
 	}
-
+  
 	&:hover {
-		border-color: #0A93DC;
-		&:before {
-			transform: scale(1);
-			opacity: 1;
-		}
+	  border-color: #0A93DC;
+	  transform: scale(1.1);
+	  transition-timing-function: ease-in-out;
+	  &:before {
+		transform: scale(1);
+		opacity: 1;
+	  }
 	}
-}
-
-.checkbox-icon {
+  }
+  
+  .checkbox-icon {
 	transition: .375s ease;
 	color: #494949;
 	svg {
-		width: 3rem;
-		height: 3rem;
+	  width: 3rem;
+	  height: 3rem;
 	}
-}
-
-.checkbox-label {
-	color: #a1a1aa;
+  }
+  
+  .checkbox-label {
+	color: #707070;
 	transition: .375s ease;
 	text-align: center;
-    font-size: 1rem;
-}
+  }
