--- react_dars_website/src/loginForm.css
@@ -1,114 +1,0 @@
-/* style for wrapper class */
-.wrapper{
-    width: 420px;
-    background: rgba(100, 100, 100, 0.6);
-    /* (Alt)
-        background: transparent; 
-        border: 2px solid rgba(255, 255, 255, .2);
-    */
-    color: #fff;
-    border-radius: 10px; 
-    padding: 30px 40px;
-}
-
-.wrapper h1{
-    font-size: 36px;
-    text-align: center;
-
-}
-
-.wrapper .input-box{
-    /* position behavior relative to its normal position in the doc*/
-    position: relative;
-    width: 100%;
-    height: 50px;
-    margin: 30px 0;
-
-}
-/* style for input-box class */
-.input-box input{
-    width: 100%;
-    height: 100%;
-    background: transparent;
-    border: 2px solid rgba(255, 255, 255, 2);
-    outline: none;
-    border-radius: 40px;
-    font-size: 16px;
-    color: #fff;
-    padding: 20px 45px 20px 20px;
-}
-
-.input-box input::placeholder{
-    color: #fff;
-
-}
-
-.input-box .icon{
-    position: absolute;
-    right: 20px;
-    top: 50%;
-    transform: translateY(-50%);
-    font-size: 16px;
-
-}
-/* style for the remember me and forgot*/
-.wrapper .remember-forgot {
-    display: flex;
-    justify-content: space-between;
-    font-size: 14.5px;
-    margin: -15px 0 15px;
-
-}
-
-.remember-forgot label input{
-    accent-color: #fff;
-    margin-right: 4px;
-
-}
-
-.remember-forgot a{
-    color: #fff;
-    /* remove the underline on hyperlink */
-    text-decoration: none;
-
-}
-/* when mouse move over, underline appear*/
-.remember-forgot a:hover{
-    /* only show underline when move over the content */
-    text-decoration: underline;
-
-}
-
-/* style for the button */
-.wrapper button{
-    width: 100%;
-    height: 45px;
-    background: #fff;
-    border: none;
-    outline: none;
-    border-radius: 40px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, .1);
-    /* change the mouse type when move to button */
-    cursor: pointer;
-    font-size: 16px;
-    color: #333;
-    font-weight: 700;
-}
- /* register link and line */
-.wrapper .register-link{
-    font-size: 14.5px;
-    text-align: center;
-    margin: 20px 0 15px;
-
-}
-
-.register-link p a{
-    color: #fff;
-    text-decoration: none;
-    font-weight: 600;
-}
-
-.register-link p a:hover{
-    text-decoration: underline;
-
-}
